third_party/WebKit/LayoutTests/fast/dom/htmlformcontrolscollection-no-img.html

<!DOCTYPE html>
<body>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<form>
<img name="imageName">
<img id="imageId">
</form>
<script>
test(function() {
    var collection = document.querySelector('form').elements;
    assert_not_exists(collection, 'imageName', 'name getter for name=');
    assert_not_exists(collection, 'imageId', 'name getter for id=');
    assert_equals(collection.namedItem('imageName'), null);
    assert_equals(collection.namedItem('imageId'), null);
    for (var i = 0; i < collection.length; ++i)
        assert_not_equals(collection[i].tagName, 'IMG');
}, 'HTMLFormControlsCollection should not handle IMG elements.');
</script>
</body>
